split client and server into separate folders
