+++ .gitignore
@@ -1,5 +1,5 @@
 node_modules
 .vscode
 .env
-/dist
-/build
+**/dist
+**/build
