+++ client/src/components/Login.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 function Login() {
   return (
     <div className="App">
