+++ client/src/components/WebPlayback.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import callSpotifyApi from '../utils/callSpotifyApi';
 
 const track = {
