+++ client/src/test/components/Login.test.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { render, screen } from '@testing-library/react';
 import Login from '../../components/Login';
