+++ client/tsconfig.json
@@ -20,7 +20,4 @@
     "noEmit": true,
     "jsx": "react-jsx"
   },
-  "include": [
-    "src"
-  ]
 }
