+++ server/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "dynamic-playlists-server",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "dependencies": {
+    "body-parser": "^1.20.2",
+    "cors": "^2.8.5",
+    "dotenv": "^16.3.1",
+    "express": "^4.18.2",
+    "node-pg-migrate": "^6.2.2",
+    "nodemon": "^3.0.1",
+    "pg": "^8.11.1",
+    "request": "^2.88.2"
+  },
+  "scripts": {
+    "start": "ts-node --esm index.ts",
+    "dev": "nodemon --ext ts,json --ignore src/__tests__ --exec ts-node --esm index.ts",
+    "migrate:up": "node-pg-migrate up",
+    "migrate:down": "node-pg-migrate down",
+    "migrate:redo": "node-pg-migrate redo",
+    "migrate:creatåe": "node-pg-migrate create"
+  },
+  "devDependencies": {
+    "@types/body-parser": "^1.19.2",
+    "@types/cors": "^2.8.13",
+    "@types/express": "^4.17.17",
+    "@types/node": "^20.4.2",
+    "@types/pg": "^8.10.2",
+    "@types/request": "^2.48.8",
+    "eslint": "^8.45.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^17.1.0",
+    "eslint-plugin-import": "^2.22.1",
+    "jest": "^26.6.3",
+    "typescript": "^4.9.5"
+  }
+}
