+++ server/routes/api/index.ts
@@ -1,7 +1,7 @@
 import { Router, Request, Response } from "express";
 
-import authRouter from "./auth";
-import usersRouter from "./users";
+import authRouter from "./auth/index.js";
+import usersRouter from "./users/index.js";
 
 const apiRouter = Router();
 apiRouter.use("/auth", authRouter);
