+++ server/routes/api/users/index.ts
@@ -1,5 +1,5 @@
 import { Router, Request, Response } from 'express';
-import { deleteUser, findUser, insertUser } from '../../../services/users';
+import { deleteUser, findUser, insertUser } from '../../../services/users/index.js';
 
 const usersRouter = Router();
