+++ server/routes/index.ts
@@ -1,5 +1,5 @@
 import { Request, Response, Router } from 'express';
-import apiRouter from './api';
+import apiRouter from './api/index.js';
 
 const router = Router();
