--- tsconfig.server.json
@@ -1,27 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2022",
-    "lib": [
-      "es6"
-    ],
-    "outDir": "./dist/server",
-    "allowJs": false,
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "isolatedModules": true,
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "noFallthroughCasesInSwitch": true,
-    "resolveJsonModule": true,
-    "skipLibCheck": true,
-    "strict": true,
-  },
-  "include": [
-    "src/server/**/*",
-  ],
-  "exclude": [
-    "node_modules",
-    "dist",
-  ]
-}
